refactor: 레시피 순서 id Long 타입으로 변경

+++ app/src/main/java/com/hyeok/recipebook/presentation/recipe/detail/step/RecipeStepCard.kt
@@ -81,7 +81,7 @@ private fun StepIcon(
 fun EditingRecipeStepCard(
     editState: RecipeStepEditState,
     modifier: Modifier = Modifier,
-    onRemoveStep: (Int) -> Unit = {}
+    onRemoveStep: (Long) -> Unit = {}
 ) {
     YorinCard(
         modifier = modifier
